Address code review feedback: fix redundant cast, add punctuation to error message
Agent-Logs-Url: https://github.com/Neotran010/volamv2/sessions/89329622-00bf-4bd9-bacd-8b9bc54f08b4

+++ VoLamV2 2/src/neo/volam2/abilities/CooldownSkills.java
@@ -93,7 +93,7 @@ public class CooldownSkills implements NeoSkill {
 	}
 	
 	public void setCooldown(Player p, double seconds) {
-		cooldownPlayers.put(p.getName(), (long) (System.currentTimeMillis() + (long)(1000*seconds)));
+		cooldownPlayers.put(p.getName(), System.currentTimeMillis() + (long)(1000 * seconds));
 	}
 
 	@Override
